refactor: Restaurant -> RestaurantItem으로 컴포넌트 이름 수정
